new footer component
refactoring and cleaning
